inline fix, commenting

+++ contracts/marble/src/marble.cpp
@@ -273,7 +273,7 @@ ACTION marble::newnft(name owner, name group_name, bool log) {
         //inline logevent
         action(permission_level{get_self(), name("active")}, get_self(), name("logevent"), make_tuple(
             "newserial"_n, //event_name
-            new_serial, //event_value
+            int64_t(new_serial), //event_value
             now, //event_time
             std::string("log new nft serial") //memo
         )).send();
